poliastro.core and thrust modules change_ecc_inc updated
