+++ src/poliastro/twobody/thrust/change_ecc_inc.py
@@ -5,14 +5,9 @@ References
 * Pollard, J. E. "Simplified Analysis of Low-Thrust Orbital Maneuvers", 2000.
 
 """
-import numpy as np
 from astropy import units as u
-from numba import njit
-from numpy import cross
-from numpy.linalg import norm
 
-from poliastro.core.elements import rv2coe
-from poliastro.core.thrust.change_ecc_inc import beta, extra_quantities
+from poliastro.core.thrust.change_ecc_inc import change_ecc_inc as change_ecc_inc_fast
 
 
 def change_ecc_inc(ss_0, ecc_f, inc_f, f):
@@ -25,50 +20,19 @@ def change_ecc_inc(ss_0, ecc_f, inc_f, f):
     ----------
     ss_0 : Orbit
         Initial orbit, containing all the information.
-    ecc_f : float
+    ecc_f : ~astropy.units.quantity.Quantity
         Final eccentricity.
-    inc_f : float
-        Final inclination.
-    f : float
-        Magnitude of constant acceleration.
+    inc_f : ~astropy.units.quantity.Quantity
+        Final inclination (rad).
+    f : ~astropy.units.quantity.Quantity
+        Magnitude of constant acceleration (km / s**2).
     """
 
-    # We fix the inertial direction at the beginning
-    ecc_0 = ss_0.ecc.value
-    if ecc_0 > 0.001:  # Arbitrary tolerance
-        ref_vec = ss_0.e_vec / ecc_0
-    else:
-        ref_vec = ss_0.r / norm(ss_0.r)
-
-    h_unit = ss_0.h_vec / norm(ss_0.h_vec)
-    thrust_unit = cross(h_unit, ref_vec) * np.sign(ecc_f - ecc_0)
-
-    inc_0 = ss_0.inc.to(u.rad).value
-    argp = ss_0.argp.to(u.rad).value
-
-    beta_0_ = beta(ecc_0, ecc_f, inc_0, inc_f, argp)
-
-    @njit
-    def a_d(t0, u_, k):
-        r = u_[:3]
-        v = u_[3:]
-        nu = rv2coe(k, r, v)[-1]
-        beta_ = beta_0_ * np.sign(
-            np.cos(nu)
-        )  # The sign of ß reverses at minor axis crossings
-
-        w_ = cross(r, v) / norm(cross(r, v))
-        accel_v = f * (np.cos(beta_) * thrust_unit + np.sin(beta_) * w_)
-        return accel_v
-
-    delta_V, beta_, t_f = extra_quantities(
-        ss_0.attractor.k.to(u.km ** 3 / u.s ** 2).value,
-        ss_0.a.to(u.km).value,
-        ecc_0,
-        ecc_f,
-        inc_0,
-        inc_f,
-        argp,
-        f,
+    a_d, delta_V, beta_, t_f = change_ecc_inc_fast(
+        ss_0 = ss_0,
+        ecc_f = ecc_f.to_value(u.one),
+        inc_f = inc_f.to_value(u.rad),
+        f = f.to_value(u.km / u.s ** 2),
     )
-    return a_d, delta_V, beta_, t_f
+
+    return a_d, delta_V, beta_, t_f * u.s
